MiniTCPTunnel_ag/requirements.txt
PySide6>=6.6.0
cryptography>=41.0.0
PyNaCl>=1.5.0
lz4>=4.3.0
qasync>=0.27.1
pydantic-settings>=2.0.0
